requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.11
protobuf==3.11.2
PyMySQL==0.9.3
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
